+++ cdk/src/stacks/control-tower.ts
@@ -3,11 +3,11 @@ import { NestedStack, NestedStackProps } from 'aws-cdk-lib';
 import { CfnInclude } from 'aws-cdk-lib/cloudformation-include';
 import { Construct } from 'constructs';
 
-export class BudgetStack extends NestedStack {
+export class ControlTowerStack extends NestedStack {
   constructor(scope: Construct, id: string, props: NestedStackProps) {
     super(scope, id, props);
     new CfnInclude(this, 'SuperwerkerTemplate', {
-      templateFile: path.join(__dirname, '..', '..', 'templates', 'budget.yaml'),
+      templateFile: path.join(__dirname, '..', '..', '..', 'templates', 'control-tower.yaml'),
     });
   }
 }
