+++ cdk/src/stacks/guardduty.ts
@@ -3,11 +3,11 @@ import { NestedStack, NestedStackProps } from 'aws-cdk-lib';
 import { CfnInclude } from 'aws-cdk-lib/cloudformation-include';
 import { Construct } from 'constructs';
 
-export class ServiceControlPoliciesStack extends NestedStack {
+export class GuardDutyStack extends NestedStack {
   constructor(scope: Construct, id: string, props: NestedStackProps) {
     super(scope, id, props);
     new CfnInclude(this, 'SuperwerkerTemplate', {
-      templateFile: path.join(__dirname, '..', '..', 'templates', 'service-control-policies.yaml'),
+      templateFile: path.join(__dirname, '..', '..', '..', 'templates', 'guardduty.yaml'),
     });
   }
 }
